Add game controller
